Move functions for load inspection to a separate module for ease of importing
